Provide PLATFORM macro for platform aware code
This is an ease of use feature, which could be done in many other ways.

The change is especially helpful if application code is run on multiple
targets but the application needs to accommodate for some platform
specifics. Another way would be to move those specifics to the platform
specific code (e.g. virt.c, tx2.c), however this forces application
specific code into the platform specific code. So allowing the application
to know on which platform it will run and make small adjustments, is IMO
cleaner.

Compilation was tested for all platforms (IMX8, ZCU, TX2, QEMU-AARCH64,
QEMU-RISCV).